Extract `fj-kernel` from `fj-app`
