+++ fj-app/Cargo.toml
@@ -13,17 +13,12 @@ categories  = ["mathematics", "rendering"]
 
 [dependencies]
 anyhow      = "1.0.56"
-approx      = "0.5.1"
 bytemuck    = "1.8.0"
 futures     = "0.3.21"
 libloading  = "0.7.2"
-map-macro   = "0.2.0"
 nalgebra    = "0.30.0"
 notify      = "5.0.0-pre.14"
-parking_lot = "0.12.0"
-parry2d-f64 = "0.8.0"
 parry3d-f64 = "0.8.0"
-spade       = "2.0.0"
 thiserror   = "1.0.30"
 threemf     = "0.2.0"
 tracing     = "0.1.32"
@@ -46,6 +41,10 @@ path    = "../fj"
 [dependencies.fj-debug]
 version = "0.5.0"
 path    = "../fj-debug"
+
+[dependencies.fj-kernel]
+version = "0.5.0"
+path    = "../fj-kernel"
 
 [dependencies.fj-math]
 version = "0.5.0"
