+++ fj-app/src/operations/circle.rs
@@ -1,11 +1,10 @@
 use fj_debug::DebugInfo;
-use fj_math::{Aabb, Point, Scalar};
-
-use crate::kernel::{
+use fj_kernel::{
     geometry::Surface,
     shape::Shape,
     topology::{edges::Cycle, faces::Face},
 };
+use fj_math::{Aabb, Point, Scalar};
 
 use super::ToShape;
