+++ fj-app/src/operations/sweep.rs
@@ -1,7 +1,6 @@
 use fj_debug::DebugInfo;
+use fj_kernel::{algorithms::sweep::sweep_shape, shape::Shape};
 use fj_math::{Aabb, Scalar, Vector};
-
-use crate::kernel::{algorithms::sweep::sweep_shape, shape::Shape};
 
 use super::ToShape;
