+++ fj-kernel/src/algorithms/mod.rs
