+++ fj-kernel/src/algorithms/triangulation.rs
@@ -2,7 +2,7 @@ use fj_math::Scalar;
 use parry2d_f64::utils::point_in_triangle::{corner_direction, Orientation};
 use spade::HasPosition;
 
-use crate::kernel::geometry;
+use crate::geometry;
 
 /// Create a Delaunay triangulation of all points
 pub fn triangulate(
