+++ fj-kernel/src/geometry/curves/circle.rs
