+++ fj-kernel/src/geometry/curves/line.rs
