+++ fj-kernel/src/geometry/curves/mod.rs
