+++ fj-kernel/src/geometry/mod.rs
