+++ fj-kernel/src/geometry/points.rs
