+++ fj-kernel/src/shape/handle.rs
