+++ fj-kernel/src/shape/iter.rs
