+++ fj-kernel/src/topology/faces.rs
@@ -8,7 +8,7 @@ use fj_math::{Aabb, Scalar, Segment, Triangle};
 use parry2d_f64::query::{Ray as Ray2, RayCast as _};
 use parry3d_f64::query::Ray as Ray3;
 
-use crate::kernel::{
+use crate::{
     algorithms::{approximation::Approximation, triangulation::triangulate},
     geometry::Surface,
     shape::Handle,
