+++ fj-kernel/src/topology/mod.rs
